fix end/start/min/max date arguments in codeblocks not respecting narrowing down to month, day or hour
